requirements.txt
pytest==7.1.2
allure-pytest==2.9.45
selene==2.0.0b5
python-dotenv==0.20.0
chromedriver_autoinstaller